Ability to override discoverable bolt URI
This introduces a minor overhaul of the discovery
service, allowing arbitrary URLs to be published to it.
That functionality is leveraged to have the enterprise
edition of the server publish a `bolt+routing` URI.

--

This has two elements that will make it confusing - the term
"discovery", and the idea of a "public" URI for Bolt.

"Discovery" in Neo could refer both to the discovery *protocol*,
used by causal clustering, or to REST API Discovery, the idea
that the HTTP API describes its own capabilities. This PR is
exclusively about REST/HTTP discovery - it is adding a capability
to the JSON document a user recieves when she does:

    GET http://localhost:7474

The "public" URI portion is more complex, and is the core of what
this PR is mucking about in. Currently, you can find a Bolt URL
to a Neo4j instance in two ways - by asking the HTTP discovery
endpoint, or by asking the routing table procedure. Currently,
both of these methods will give you URLs ultimately originating
from the configuration of the `advertised_address` on the bolt
connector.

However, they serve two different purposes. The routing table is
used *internally* by Bolt Drivers, providing addresses to connect
to exact and specific Neo4j instances.

The Discovery URL is used *externally*, as the root URL to give to
a driver. By connecting via HTTP and retrieving the discovery
document, tools like the browser find out the appropriate URL
to pass when creating a Bolt Driver.

Hence - a sensible "Advertised" URL is always a "bolt://..." URL
pointing to the specific instance being configured. However, a
sensible "Discoverable" URL is sometimes different - specifically,
it may very reasonably be "bolt+routing://path-to-some-load-balancer".